Taken care of wishes 2, and 3 I posted about in the devel list.
(2) Fast close of QDevelop
Now, qdevelop will save the project configuration while the window is 
hidden. Now QDevelop will "look" like it's been close in no time, however
some operations (the saving of the sqlite DB) will be done while the 
main window is hidden.

Look into MainImpl::closeEvent() for more details.


(3) All project related operations are in the project menu
Noe the project menu is more populated, and I moved all the entries
from the file menu which had to do with projects to this menu.

I added a new action which gives you an open project dialog. This
was a fork of the regular open file dialog... which means that the
dir variable has been moved 3 lines above, and now is a global static
variable (shared between 2 functions). Should we move it to the class 
instead...?

Look in line 916 for more details.

The moving of menus was done by openning the main.ui file in kate and 
moving arround xml lines... is there a smarter way...? it feels like a 
dirty hack