This fixed wish number (4).
The tools configuration dialog, has an option to disable the check on applicaiton
startup. This can be used by people who do not have gdb or ctags for example.

I also fixed the directory delimiter on Windows, and removed an exra line in the open
project dialog. Some minor formatiing fixes were applied on the source.