Nodifications to the subclassing dialog:
Now actions are displayed in that dialog as well. You also have a input line for filterring the 
tree view. The input line is based on regexp and not exact (string) matching, so you might
want to toy with it.

Added two new images, stollen from KDE, so those are free to distribute.

Closes issue #30.